requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
albumentations>=1.3.1
tqdm>=4.65.0
PyYAML>=6.0.1
pandas>=1.3.0
requests>=2.26.0
opencv-python>=4.8.0  # Full version for GUI support 